Update bash script
Initialize string variables for better comparison of android versions.

+++ bin/unin.sh
@@ -1,5 +1,7 @@
 #!/data/local/tmp/recovery/busybox sh
+
 BUSYBOX=/data/local/tmp/recovery/busybox
+
 ${BUSYBOX} mount -o remount,rw /system
 
 # Remove cwm
@@ -22,8 +24,17 @@ if [ -e /system/bin/recovery.cpio ]; then
     ${BUSYBOX} rm /system/bin/recovery.cpio
 fi
 
-${BUSYBOX} rm /system/bin/recovery.sh
+# Remove byeselinux lkm
+if [ -f "/system/lib/modules/byeselinux.ko" ]; then
+    ${BUSYBOX} rm /system/lib/modules/byeselinux.ko
+fi
 
+# Remove recovery script from system
+if [ -f "/system/bin/recovery.sh" ]; then
+    ${BUSYBOX} rm /system/bin/recovery.sh
+fi
+
+# Restore chargemon & e2fsck binary files
 if [ -e /system/bin/chargemon.bin ]; then
    ${BUSYBOX} mv /system/bin/chargemon.bin /system/bin/chargemon
 fi
